rename top-level package simply 'pathops'
Initially I thought of making this only the bindings for the skia
pathops module, and have a separate higher-level package like
booleanOperations that would use that and provide a UFO PointPen
interface for UFO editors.
That's why I called the top-level module "skia", with submodules like
core and pathops following skia's API.

However, after discussing with Adrien, we decided to keep this simpler
and have a single "pathops" python package which contains a cython
extension module calling the skia C++ library, and which also provides
a python interface similar to booleanOperations: one function per
each boolean operation, taking a list of UFO contours (with drawPoints
method) and an output pointPen to draw the result on.
This way, we can call this skia-pathops directly from fontmake or TruFont.

BooleanOperations (the python module) can still use skia-pathops
internally, as it provides other useful classes like the BooleanGlyph
which we don't currently need in our pipeline. But that's a separate
concern.

The pathops/__init__.py is where the python api will be exported.
Then we have a pathops._pathops.so extension module (built from the
pathops/_pathops.pyx cython file) that translates between Python and
C++.

Inside pathops/_skia/ folder are the *.pxd files with the external
declarations for the skia library that are called from Cython code.

I added the _ to the _skia subfolder to mean this it is private.
i.e. this is not a full-blown python bindings to skia, just the
headers we care about here.

I hope this makes sense.