[_pathops.pyx] handle endPath and optionally raise with open paths
From my tests, it appears the SkOpBuilder silently auto-closes paths
that are sent to it.
We want to retain the same behavior as booleanOperations and raise
when operating on open UFO contours, so that the caller has to
decide how to handle them (whether to skip them or to auto-close
them).