build-requirements.txt: don't require the most recent setuptools
we are implementing everything ourselves in custom build_ext and
build_clib commands (as we need to pass compiler specific options)
so we actually don't need such a recent setuptools.

The current python distributions come with 28.8.
I tried to build with an even oldder one that Frederik had issues
with (18.4 from more than two years ago) it builds without problems.