expose svg-like SkPath::arcTo method as Path.arcTo()
https://skia.org/user/api/SkPath_Reference#SkPath_arcTo

The SkPath::arcTo method implements the functionality of SVG arcs. The only difference is that
the SVG 'sweep-flag' value is opposite the integer value of 'sweep' argument in SkPath::arcTo.
That is, SVG 'sweep-flag' uses 1 for clockwise, whereas SkPathDirection::kCW cast to int is 0.

So in nanosvg, if we want to use this to draw SVG arc commands onto a pathops.Path, we need to
make sure that we negate the sweep flag.